Somewhat working state

+++ Assets/Scripts/Base/Runtime/Management/LevelSpawner/B_LC_LevelManager.cs
@@ -4,12 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using UnityEngine;
 
-namespace Base
-{
+namespace Base {
     public enum LevelType { Tutorial, Main }
 
-    public class B_LC_LevelManager : B_M_ManagerBase
-    {
+    public class B_LC_LevelManager : B_M_ManagerBase {
         public static B_LC_LevelManager instance;
 
         public Action<int> OnLevelChangedAction;
@@ -26,16 +24,13 @@ namespace Base
         [HideInInspector] public Transform LevelHolder { get; private set; }
         public static Transform ObjectSpawnParent;
 
-        private int tutorialPlayed
-        {
-            get
-            {
-                return /*B_GM_GameManager.instance.Save.TutorialPlayed*/ (int)SaveSystem.GetDataInt(Enum_Saves.Save_1, Enum_Save_1.TutorialPlayed);
+        private int tutorialPlayed {
+            get {
+                return /*B_GM_GameManager.instance.Save.TutorialPlayed*/ SaveSystem.GetDataInt(Enum_Saves.Save_1, Enum_Save_1.TutorialPlayed);
             }
         }
 
-        public override Task ManagerStrapping()
-        {
+        public override Task ManagerStrapping() {
             if (instance == null) instance = this; else Destroy(this.gameObject);
 
             LevelHolder = GameObject.Find("LevelHolder").GetComponent<Transform>();
@@ -58,15 +53,12 @@ namespace Base
             return base.ManagerStrapping();
         }
 
-        public override Task ManagerDataFlush()
-        {
+        public override Task ManagerDataFlush() {
             return base.ManagerDataFlush();
         }
 
-        public void LoadInLevel(int levelNumber)
-        {
-            switch (tutorialPlayed)
-            {
+        public void LoadInLevel(int levelNumber) {
+            switch (tutorialPlayed) {
                 case 0:
                     if (levelNumber >= TutorialLevels.Count) levelNumber = 0;
                     InitateNewLevel(TutorialLevels[levelNumber]);
@@ -79,10 +71,8 @@ namespace Base
             }
         }
 
-        public void LoadInNextLevel()
-        {
-            switch ((int)SaveSystem.GetDataInt(Enum_Saves.Save_1, Enum_Save_1.GameFinished))
-            {
+        public void LoadInNextLevel() {
+            switch ((int)SaveSystem.GetDataInt(Enum_Saves.Save_1, Enum_Save_1.GameFinished)) {
                 case 0:
                     InitateNewLevel(LevelToLoad());
                     break;
@@ -93,19 +83,16 @@ namespace Base
             }
         }
 
-        public void ReloadCurrentLevel()
-        {
+        public void ReloadCurrentLevel() {
             InitateNewLevel(currentLevel);
         }
 
-        private void InitateNewLevel(GameObject levelToInit)
-        {
+        private void InitateNewLevel(GameObject levelToInit) {
             B_CES_CentralEventSystem.OnBeforeLevelLoaded.InvokeEvent();
             if (CurrentLevel != null) { Destroy(CurrentLevel); CurrentLevel = null; currentLevel = null; }
             CurrentLevel = GameObject.Instantiate(levelToInit, LevelHolder);
             currentLevel = levelToInit;
-            switch (tutorialPlayed)
-            {
+            switch (tutorialPlayed) {
                 case 0:
                     CurrentLevelIndex = Array.IndexOf(TutorialLevels.ToArray(), levelToInit);
                     SaveSystem.SetData(Enum_Saves.Save_1, Enum_Save_1.PlayerLevel, CurrentLevelIndex);
@@ -128,23 +115,20 @@ namespace Base
             ObjectSpawnParent = LevelHolder.GetChild(0);
         }
 
-        private GameObject LevelToLoad()
-        {
-            switch (tutorialPlayed)
-            {
+        private GameObject LevelToLoad() {
+            switch (tutorialPlayed) {
                 case 0:
-                    if (CurrentLevelIndex + 1 >= TutorialLevels.Count)
-                    {
+                    if (CurrentLevelIndex + 1 >= TutorialLevels.Count) {
                         CurrentLevelIndex = 0;
                         SaveSystem.SetData(Enum_Saves.Save_1, Enum_Save_1.TutorialPlayed, 1);
+                        //SaveSystem.SetData(Enum_Saves.Save_1, Enum_Save_1.TutorialPlayed, );
                         //B_GM_GameManager.instance.Save.TutorialPlayed = 1;
                         return MainLevels[0];
                     }
                     return TutorialLevels[CurrentLevelIndex + 1];
 
                 case 1:
-                    if (CurrentLevelIndex + 1 >= MainLevels.Count)
-                    {
+                    if (CurrentLevelIndex + 1 >= MainLevels.Count) {
                         SaveSystem.SetData(Enum_Saves.Save_1, Enum_Save_1.GameFinished, 1);
                         //B_GM_GameManager.instance.Save.GameFinished = 1;
                         return RandomSelectedLevel();
@@ -154,35 +138,29 @@ namespace Base
             return null;
         }
 
-        private void CheckLevels()
-        {
-            for (int i = 0; i < MainLevels.Count; i++)
-            {
+        private void CheckLevels() {
+            for (int i = 0; i < MainLevels.Count; i++) {
                 Debug.Log(MainLevels[i].name);
             }
             Debug.Log("//----------//");
-            for (int i = 0; i < TutorialLevels.Count; i++)
-            {
+            for (int i = 0; i < TutorialLevels.Count; i++) {
                 Debug.Log(TutorialLevels[i].name);
             }
         }
 
-        private GameObject RandomSelectedLevel()
-        {
+        private GameObject RandomSelectedLevel() {
             if (MainLevels.Count <= 1) { return MainLevels[0]; }
             GameObject obj = MainLevels[UnityEngine.Random.Range(0, MainLevels.Count)];
             if (currentLevel == obj) return RandomSelectedLevel();
             return obj;
         }
 
-        private void SaveOnNextLevel()
-        {
+        private void SaveOnNextLevel() {
             SaveSystem.SetData(Enum_Saves.Save_1, Enum_Save_1.PreviewLevel, PreviewLevelIndex + 1);
             //B_GM_GameManager.instance.Save.PreviewLevel = PreviewLevelIndex + 1;
         }
 
-        private void OnDestroy()
-        {
+        private void OnDestroy() {
             instance = null;
             ObjectSpawnParent = null;
         }
